home_node/backend/public_api/requirements.txt
aiohttp
fastapi
passlib[bcrypt]
python-jose[cryptography]
python-multipart
redis
ujson
uvicorn
